Move some util funcs for testing to export_tests.go
